[64-bit] General Development
Have stripped USPi out for the moment as, despite initial work to get it 64-bit compatible, there was a bigger problem.
This was the vector table (IRQ/FIQ/Sync/SError) handling code. Although it was perceived to be working earlier, it looks
like it actually wasn't.

At this point, the 64-bit build is running and shows this in debug mode:

_kernel_swi_handler : 00000007
*_kernel_swi_handler : 00000009

It then "hangs". This makes sense so far, as the CLI userspace program calls SWI 0x9 to OS_PrintString ("*") followed by
an SWI call 0x7 to OS_ReadC. This is (without checking) going to be busy waiting, since the USPi stuff has been knocked out
of the build currently, and since the sync_handler starts off with DAIF implying no-interrupts-are-on, what's presented on-screen
appears to be a "frozen" machine, which would make some sense.

A lot more work to go on here, but this is - odd as it may seem - progress this evening.

+++ kernel/c/common/include/swi.h
@@ -31,7 +31,7 @@
 	};
 
 	// internal SWI handlers
-	unsigned int _kernel_swi_handler( uintptr_t swi );
+	uintptr_t _kernel_swi_handler( uintptr_t swi );
 
 	unsigned int _kernel_swi_os_setmode( void );
 	unsigned int _kernel_swi_os_setcolour( void );
